+++ docs/CHANGELOG.rst
@@ -26,7 +26,8 @@ Changed
 
 Fixed
 -----
-
+- Fix ``qorts-qc`` process so it can be used with hg19 annotation
+  and improve error handling when files are missing
 
 ===================
 55.0.0 - 2023-11-13